Minh branch

+++ src/main/java/library/AccountRepository.java
@@ -1,13 +1,14 @@
 package library;
 
+import library.entity.Account;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-public interface AccountRepository extends JpaRepository<Account, Long> {
+public interface AccountRepository extends JpaRepository<Account, Integer> {
     boolean existsByUsername(String username);
 
     Account findAccountByUsername(String username);
 
     @Query("SELECT MAX(a.id) FROM Account a")
-    Long findMaxId();
+    Integer findMaxId();
 }
